hamcrest library removed, Refractor class name and test added
